handle error logging outside of cobra
